refactor: split __main__.py into separate modules
- auth.py: verify_admin, now_iso
- helpers.py: get_session_state, get_session_last_activity, serialize_cached_message
- ws_manager.py: ConnectionManager class + global instance
- broadcast.py: WebSocket broadcast loop + collect_full_status
- routers/sessions.py: session REST endpoints using APIRouter
- routers/notes.py: notes CRUD endpoints using APIRouter
- routers/ego.py: EGO endpoints using APIRouter
- __main__.py: minimal — only imports, router registration, WS endpoint, cleanup scheduler
