+++ src/plugins/nonebot_plugin_chat_monitor/routers/__init__.py
@@ -1,0 +1,1 @@
+"""Chat Monitor 路由模块"""
